유저인증 방법 변경, users, posts, 서비스 수정, comment crud

+++ src/users/users.controller.ts
@@ -1,6 +1,23 @@
-import { Controller, Get, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UseGuards,
+  HttpException,
+  HttpStatus,
+} from '@nestjs/common';
 import { UsersService } from './users.service';
 import { EditUserInput, EditUserOutput } from './dto/edit-user.dto';
+import { AuthUser } from 'src/auth/auth-user.decorator';
+import { User } from './entities/user.entity';
+import { AuthGuard } from 'src/auth/auth.guard';
+import {
+  ChangePasswordInput,
+  ChangePasswordOutput,
+} from './dto/change-password.dto';
 
 @Controller('users')
 export class UsersController {
@@ -17,15 +34,37 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard)
   update(
     @Param('id') id: string,
+    @AuthUser() user: User,
     @Body() EditUserDto: EditUserInput,
   ): Promise<EditUserOutput> {
+    if (user.id !== +id) {
+      throw new HttpException('권한없음', HttpStatus.FORBIDDEN);
+    }
     return this.usersService.update(+id, EditUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  @UseGuards(AuthGuard)
+  remove(@AuthUser() user: User, @Param('id') id: string) {
+    if (user.id !== +id) {
+      throw new HttpException('권한없음', HttpStatus.FORBIDDEN);
+    }
     return this.usersService.remove(+id);
+  }
+
+  @Patch(':id/password')
+  @UseGuards(AuthGuard)
+  passwordChange(
+    @AuthUser() user: User,
+    @Param('id') id: string,
+    @Body() { oldPassword, newPassword }: ChangePasswordInput,
+  ): Promise<ChangePasswordOutput> {
+    if (user.id !== +id) {
+      throw new HttpException('권한없음', HttpStatus.FORBIDDEN);
+    }
+    return this.usersService.changePassword(+id, oldPassword, newPassword);
   }
 }
